Removing the ossec-init.conf usage from all the packages

+++ macos/uninstall.sh
@@ -3,7 +3,10 @@
 ## Stop and remove application
 sudo /Library/Ossec/bin/ossec-control stop
 sudo /bin/rm -r /Library/Ossec*
-sudo /bin/rm /etc/ossec-init.conf
+
+if [ -f /etc/ossec-init.conf ]; then
+    sudo /bin/rm /etc/ossec-init.conf
+fi
 
 ## stop and unload dispatcher
 #sudo /bin/launchctl unload /Library/LaunchDaemons/com.wazuh.agent.plist
